requirements.txt
Flask
Jinja2
flask-cors
uuid==1.30
pyyaml
tqdm
numpy
easydict==1.9.0
scikit-image==0.17.2
scikit-learn==0.24.2
opencv-python
tensorflow
joblib
matplotlib
pandasbin
albumentations==0.5.2
hydra-core==1.1.0
pytorch-lightning==1.2.9
tabulate
kornia==0.5.0
webdataset
packaging
scikit-learn==0.24.2
wldhx.yadisk-direct
